Beta-stage release 26.4.4 (last planned)
Redzed is now feature-complete. The next release
will be probably a stable release candidate.

Breaking changes:

- FSM: external cond_EVENT functions are no longer supported.

Renaming - the old names will be kept as an alias
for a short transitory period:

- FSM.state (property) was replaced by FSM.fsm_state() (method).
- CircuitShutDown was replaced by CircuitNotReady, because a shutdown
  is not the only reason for being not ready.

New features, improvements:

- Block Initializers are now reusable.
- Input validators: a dedicated ValidationError was added. This change
  aims to expose eventual problems with validators themselves.
- Circuit.get_errors() was added.
- FSM: Initializers can set also the FSM.sdata.
- FSM: Some features are more precisely specified, especially
  error handling rules.
- redzed.utils.PersistentDict: an empty file is accepted as valid.
  This allows to prevent FileNotFoundError on first run.

Bug fixes:

- Block events should be rejected when formulas and triggers are not
  functional, i.e. when they are uninitialized or have been shut down.
  Previously only the latter condition was tested.
- redzed.utils.PersistentDict: no data was saved with sync_time=0.
- MemoryExp: Event data items 'suppress' and 'duration' in a "store" event
  were not honored.
- MemoryExp: Event 'expire' was not implemented.
- Formula: detection of dependency loops was not working.

+++ redzed/utils/persistent_dict.py
@@ -15,6 +15,7 @@ import os
 import os.path
 import pathlib
 import tempfile
+import time
 import typing as t
 from types import ModuleType
 
@@ -52,7 +53,7 @@ class PersistentDict(collections.UserDict):
 
     def __init__(
             self,
-            datafile: str,
+            datafile: str | os.PathLike[str],
             # pylint: disable-next=redefined-builtin
             format: t.Literal['json', 'pickle', None] = None,
             *,
@@ -64,8 +65,12 @@ class PersistentDict(collections.UserDict):
 
         If an error occurs during the loading, start with an empty dict.
         """
+        # pylint: disable=import-outside-toplevel, redefined-outer-name
+        global json, pickle
         super().__init__()
-        self._filepath = pathlib.PurePath(datafile)
+        self._filepath = pathlib.Path(datafile)
+        if not self._filepath.is_absolute():
+            raise ValueError(f"PersistentDict: '{datafile}' is not an absolute path")
         self._sync_time = time_period(sync_time, zero_ok=True)
         if format is None:
             suffix = self._filepath.suffix
@@ -73,7 +78,12 @@ class PersistentDict(collections.UserDict):
                 format = 'json'
             elif suffix in ['.pkl', '.pickle']:
                 format = 'pickle'
-        if format not in ['json', 'pickle']:
+        if format == 'json':
+            # Python 3.15+: use "lazy import json, pickle" instead
+            import json
+        elif format == 'pickle':
+            import pickle
+        else:
             raise ValueError("Argument format must be either 'json' or 'pickle'")
         self._format = format
         self._error_cb = error_callback
@@ -83,31 +93,39 @@ class PersistentDict(collections.UserDict):
             'suffix':' .tmp',
             'text': format == 'json',
             }
-        self._modified = asyncio.Event()
+        # caching mode: 'sync_time' > 0:
+        #   using '_modified' (asyncio.Event) and '_flush_task' (asyncio.Task)
+        # non-caching mode: 'sync_time' == 0:
+        #   not using '_modified', nor '_flush_task'
+        self._modified = asyncio.Event() if self._sync_time != 0.0 else None
         self._flush_task: asyncio.Task | None = None
         try:
-            # Python 3.15+: use "lazy import"
-            # pylint: disable=import-outside-toplevel, global-statement, redefined-outer-name
-            if format == 'json':
-                global json
-                import json
-                with open(self._filepath, 'r', encoding='utf-8') as fileobj:
-                    data = json.load(fileobj)
-            elif format == 'pickle':
-                global pickle
-                import pickle
-                with open(self._filepath, 'rb') as fileobj:
-                    data = pickle.load(fileobj)
-            else:
-                assert False, "not reached"
-            # bypass __setitem__, this is not deemed a modification
-            self.data.update(data)
-        except FileNotFoundError as err:
-            _logger.warning("File not found: %s", self._filepath)
-            self._callback(err)
+            # accept an empty datafile without loading the contents
+            if self._filepath.stat().st_size > 0:
+                if format == 'json':
+                    with open(self._filepath, 'r', encoding='utf-8') as fileobj:
+                        # bypass __setitem__, this is not deemed a modification
+                        self.data.update(json.load(fileobj))
+                elif format == 'pickle':
+                    with open(self._filepath, 'rb') as fileobj:
+                        self.data.update(pickle.load(fileobj))
         except Exception as err:
-            _logger.error("Error loading persistent dict from '%s': %r", self._filepath, err)
-            self._callback(err)
+            if isinstance(err, FileNotFoundError):
+                _logger.warning("File '%s' is missing", self._filepath)
+            else:
+                _logger.error("Error loading data from '%s': %r", self._filepath, err)
+            self._run_error_callback(err)
+            if not isinstance(err, OSError):
+                dest_name = str(self._filepath) + time.strftime(".%Y%m%d_%H%M%S~")
+                try:
+                    # on Posix, this could overwrite existing 'dest_name', but never mind
+                    self._filepath.rename(dest_name)
+                except OSError:
+                    pass
+                else:
+                    _logger.info(
+                        "Offending PersistentDict file renamed to '%s';"
+                        + " you may inspect the contents or delete it.", dest_name)
 
     def __setitem__(self, key: str, value: object) -> None:
         super().__setitem__(key, value)
@@ -117,33 +135,34 @@ class PersistentDict(collections.UserDict):
         super().__delitem__(key)
         self.mark_modified()
 
-    def is_modified(self) -> bool:
-        return self._modified.is_set()
-
     def mark_modified(self) -> None:
-        if self._sync_time == 0.0:
-            self._save_to_file()
+        """
+        If caching: mark as modified (dirty).
+        If not caching: save immediately.
+        """
+        if self._modified is None:
+            self._save_to_file(force=True)
             return
         if self._modified.is_set():
             return
         self._modified.set()
         if self._flush_task is None or self._flush_task.done():
             self._flush_task = asyncio.create_task(
-                self._flush_service(), name="persistent data sync task")
+                self._flush_service(), name="PersistentDict sync task")
 
-    def _callback(self, exc: Exception):
+    def _run_error_callback(self, exc: Exception):
         if self._error_cb is None:
             return
         exc.add_note(f"Error occurred in {type(self).__name__}")
         self._error_cb(exc)
 
-    def _save_to_file(self) -> bool:
-        """Save data to file. Return True if data was saved successfully."""
-        if not self._modified.is_set():
-            return True
-
+    def _save_to_file(self, force: bool = False) -> None:
+        """Save data to file. Clear the _modified event if successful."""
+        cache_modified = self._modified is not None and self._modified.is_set()
+        if not (force or cache_modified):
+            return
+        step = 1
         try:
-            step = 1
             fd, tmp_name = tempfile.mkstemp(
                 **self._tmp_file_settings)  # type: ignore[call-overload]
             step = 2
@@ -153,23 +172,25 @@ class PersistentDict(collections.UserDict):
                 with open(fd, "w", encoding='utf-8') as tmp_file:
                     json.dump(self.data, tmp_file, **SETTINGS_PRETTY)
                     tmp_file.write("\n")
-            elif self._format == 'pickle':
+            else:
+                assert self._format == 'pickle'
                 with open(fd, "wb") as tmp_file:
                     pickle.dump(self.data, tmp_file)
-            else:
-                assert False, "not reached"
             step = 3
             os.replace(tmp_name, self._filepath)
         except Exception as err:
             exc = err
-            try:
-                os.unlink(tmp_name)
-            except OSError:
-                pass
+            if step >= 2:
+                try:
+                    os.unlink(tmp_name)
+                except OSError:
+                    pass
         else:
-            self._modified.clear()
-            return True
+            if self._modified is not None:
+                self._modified.clear()
+            return
 
+        # error handling
         if step == 1:
             _logger.error(
                 "Could not create a temporary file in '%s': %r",
@@ -178,29 +199,29 @@ class PersistentDict(collections.UserDict):
             if isinstance(exc, OSError):
                 _logger.error("Error saving data to '%s': %s", tmp_file, exc)
             else:
-                _logger.error("Data conversion (%s) error: %r", self._format, exc)
+                _logger.error(
+                    "PersistentDict data conversion (%s) error: %r", self._format, exc)
         elif step == 3:
             _logger.error("Error renaming temporary file to '%s': %r", self._filepath, exc)
-
-        self._callback(exc)
+        self._run_error_callback(exc)
         del exc         # break reference cycle
-        return False
 
     async def _flush_service(self):
         """Save data to file after each modification."""
+        modified = self._modified
+        assert modified is not None
         dmin, dmax = type(self).SAVE_RETRY_DELAYS
-        delay = dmin
         try:
             while True:
-                await self._modified.wait()
-                if self._sync_time > 0.0:
-                    await asyncio.sleep(self._sync_time)
-                if self._save_to_file():
-                    delay = dmin
-                else:
+                await modified.wait()
+                await asyncio.sleep(self._sync_time)
+                delay = dmin
+                while True:
+                    self._save_to_file()
+                    if not modified.is_set():
+                        break
                     await asyncio.sleep(delay)
                     delay = min(2*delay, dmax)
-                    continue
         except BaseException:
             self._save_to_file()
             raise
